[grafana] Apply lint-review cleanups
Steady-state wording for the SMTP-credential comments, drop a consumer
cross-reference from the iris health docstring, hoist the crashloop scope
labels to constants, and share the test BridgeConfig via conftest.
