+++ infra/grafana/provisioning/alerting/contact-points.yaml
@@ -8,9 +8,8 @@
 # ops-critical carries two receivers so a critical alert reaches email and Slack
 # without relying on routing-tree subtleties; ops-slack is the warning tier.
 # $SLACK_ALERTS_WEBHOOK is expanded from the container environment (Secret
-# Manager: marin-grafana-slack-webhook). Email delivery additionally needs the
-# SMTP credentials secret — until a Workspace admin fills it, email is inert
-# while Slack still delivers.
+# Manager: marin-grafana-slack-webhook). Email delivery additionally needs valid
+# SMTP credentials — without them email is inert while Slack still delivers.
 apiVersion: 1
 
 contactPoints:
