+++ infra/grafana/tests/test_bridge.py
@@ -9,7 +9,8 @@ from datetime import UTC, datetime
 
 import pyarrow as pa
 from cache import TtlCache
-from config import BridgeConfig, ClusterTarget
+from config import ClusterTarget
+from conftest import bridge_config
 from finelog.errors import QueryResultTooLargeError
 from github_source import GithubSource
 from k8s_source import K8sFleet
@@ -22,20 +23,6 @@ TO_MS = FROM_MS + 3_600_000
 MARIN = ClusterTarget(
     name="marin", project="p", zone="z", instance_filter="name = finelog-marin", controller_filter="labels.x=true"
 )
-
-
-def bridge_config(cache_ttl: float = 20.0) -> BridgeConfig:
-    return BridgeConfig(
-        max_rows=1000,
-        cache_ttl=cache_ttl,
-        query_timeout_ms=5000,
-        iris_cache_ttl=15.0,
-        github_cache_ttl=60.0,
-        k8s_cache_ttl=30.0,
-        http_timeout=5.0,
-        github_token=None,
-        cw_read_token=None,
-    )
 
 
 def finelog_result(**columns: list) -> pa.Table:
